refactor: Enum 타입 엔티티에 속하도록 변경
