--- src/main/java/com/clover/habbittracker/domain/emoji/entity/Domain.java
@@ -1,13 +1,0 @@
-package com.clover.habbittracker.domain.emoji.entity;
-
-public enum Domain {
-	COMMENT, POST;
-
-	public boolean isPost() {
-		return this == POST;
-	}
-
-	public boolean isComment() {
-		return this == COMMENT;
-	}
-}
