--- src/main/java/com/clover/habbittracker/domain/emoji/entity/Type.java
@@ -1,5 +1,0 @@
-package com.clover.habbittracker.domain.emoji.entity;
-
-public enum Type {
-	SMILE, ANGRY, SAD, SURPRISED, NONE
-}
